server/requirements.txt
librosa
mcp
numpy
pygame-ce
python-dotenv
requests
google-genai